scoring_service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
tritonclient[http]==2.44.0
psycopg2-binary==2.9.9
numpy==1.26.4
prometheus-client==0.20.0